backend/requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio>=3.7.1,<5.0.0
attrs==25.3.0
certifi==2025.8.3
click==8.3.0
databricks-sdk>=0.20.0
dataclasses-json==0.6.7
deepgram-sdk==4.8.1
deprecation==2.1.0
distro==1.9.0
fastapi>=0.104.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jiter==0.11.0
marshmallow==3.26.1
multidict==6.6.4
mypy-extensions==1.1.0
openai==1.109.0
packaging==25.0
psycopg2-binary==2.9.10
replicate==1.0.7
propcache==0.3.2
pydantic==2.11.9
pydantic-core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.6
pyyaml==6.0.2
sniffio==1.3.1
starlette==0.27.0
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
